+++ noodles-bam/src/bai/async/io.rs
@@ -1,0 +1,5 @@
+//! Async BAI I/O.
+
+mod reader;
+
+pub use self::reader::Reader;